refactor: mise à jour des routes API pour products et admin

+++ routes/api.php
@@ -1,9 +1,11 @@
 <?php
 
+use App\Http\Controllers\API\AdminController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\API\AuthController;
 use App\Http\Controllers\API\CategoryController;
+use App\Http\Controllers\API\ProductController;
 
 Route::get('/user', function (Request $request) {
     return $request->user();
@@ -15,9 +17,10 @@ Route::post('/register', [AuthController::class, 'register']);
 Route::post('/login', [AuthController::class, 'login']);
 Route::get('/categories', [CategoryController::class, 'index']);
 Route::get('/categories/{id}', [CategoryController::class, 'show']);
+Route::get('/products', [ProductController::class, 'index']);
+Route::get('/products/slug/{slug}', [ProductController::class, 'show']);
 
-
- Route::middleware('auth:sanctum')->group(function () {
+Route::middleware('auth:sanctum')->group(function () {
     
     Route::post('/logout', [AuthController::class, 'logout']);
     
@@ -26,7 +29,17 @@ Route::get('/categories/{id}', [CategoryController::class, 'show']);
 
 
 Route::middleware(['auth:sanctum', 'admin'])->group(function () {
-    Route::post('/categories', [CategoryController::class, 'store']);
-    Route::put('/categories/{id}', [CategoryController::class, 'update']);
-    Route::delete('/categories/{id}', [CategoryController::class, 'destroy']);
+
+    Route::get('/admin/users',[AdminController::class,'getAllUsers']);
+    Route::get('/admin/users/{id}',[AdminController::class,'getUserById']);
+    Route::put('/admin/users/{id}/render-employee',[AdminController::class,'renderUserEmployee']);
+
+
+    Route::post('/admin/categories', [CategoryController::class, 'store']);
+    Route::put('/admin/categories/{id}', [CategoryController::class, 'update']);
+    Route::delete('/admin/categories/{id}', [CategoryController::class, 'destroy']);
+
+    Route::post('/admin/products', [ProductController::class, 'store']);
+    Route::put('/admin/products/{id}', [ProductController::class, 'update']);
+    Route::delete('/admin/products/{id}', [ProductController::class, 'destroy']);
 });
